Filter out snapshots from squash diff
Using git format-patch to prepare the filtered diff on fly and
pass it to the git am.
Simulate the git cherry-pick behavior to save the commit message.

Seems like git am doesn't throw an error if diff is empty.

git diff won't work here as we would loose the message.